BACKAppLiv/src/main/java/com/example/BACKAppLiv/dto/ProductDto.java
package com.example.BACKAppLiv.dto;

public class ProductDto {
    private String name;
    private Long price;

    private String category;

    private String image;

    public ProductDto(String name, Long price, String category, String image) {
        this.name = name;
        this.price = price;
        this.category = category;
        this.image = image;
    }

    public ProductDto() {

    }




// Getters et setters

    public String getImage() {
        return image;
    }

    public void setImage(String image) {
        this.image = image;
    }

    public Long getPrice() {
        return price;
    }

    public void setPrice(Long price) {
        this.price = price;
    }

    public String getCategory() {
        return category;
    }

    public void setCategory(String category) {
        this.category = category;
    }

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }
}

